Restore logging of MarkLogic XQuery in phase results
Also cleans up a bit more ML fs code that is no longer used.
